+++ src/app/modules/analyst/components/report/malfunction-report/malfunction-report.component.html
@@ -1,26 +1,20 @@
 <mat-form-field>
-    <input matInput (keyup)="applyFilter($event.target.value)" placeholder="Filter">
-  </mat-form-field>
+  <input matInput (keyup)="applyFilter($event.target.value)" placeholder="Filter">
+</mat-form-field>
   
-  <div class="mat-elevation-z8">
-    <table mat-table [dataSource]="dataSource" matSort>
+<div class="mat-elevation-z8">
+  <table mat-table [dataSource]="dataSource" matSort>
+
+    <ng-container *ngFor="let columnName of displayedColumns; let i = index" [matColumnDef] = "columnName">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header> {{columnName}} </th>
+      <td mat-cell *matCellDef="let row"> {{row[columnName]}} </td>
+    </ng-container>
   
-      <!--
-      <ng-container matColumnDef="malfunction">
-        <th mat-header-cell *matHeaderCellDef mat-sort-header> Несправність </th>
-        <td mat-cell *matCellDef="let row"> {{row.fieldName}} </td>
-      </ng-container>
-      -->
-      <ng-container *ngFor="let columnName of displayedColumns; let i = index" [matColumnDef] = "columnName">
-        <th mat-header-cell *matHeaderCellDef mat-sort-header> {{columnName}} </th>
-        <td mat-cell *matCellDef="let row"> {{row[columnName]}} </td>
-      </ng-container>
+    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+  </table>
   
-      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-    </table>
-  
-    <mat-paginator [pageSizeOptions]="[5, 10, 25, 100]"></mat-paginator>
-  </div>
+  <mat-paginator [pageSizeOptions]="[5, 10, 25, 100]"></mat-paginator>
+</div>
