+++ src/app/modules/analyst/components/report/malfunction-report/malfunction-report.component.ts
@@ -3,48 +3,9 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { VehicleTypeService } from 'src/app/modules/shared/services/vehicle-type.service';
-import { StatisticsService } from 'src/app/modules/shared/services/statistics.service';
+import { StatisticsService, CreateMatTableRowFromStatistics } from 'src/app/modules/shared/services/statistics.service';
 import { Statistics } from 'src/app/modules/shared/models/statistics';
 import { VehicleType } from 'src/app/modules/shared/models/vehicleType';
-
-function CreateMatTableRowFromStatistics(statistics: Statistics, columns: string[]): {}
-{
-  let dict = {};
-  let i = 0;
-  dict[columns[i]] = statistics.fieldName;
-  statistics.statistics.forEach(num => {
-    i += 1;
-    dict[columns[i]] = num;
-  });
-
-  return dict;
-}
-
-const MY_ROWS: Statistics[] = [
-  {
-    fieldName: "Зламані мої поручні",
-    statistics: [1,2,3,4]
-  },
-  {
-    fieldName: "Зникли мої поручні",
-    statistics: [4,3,2,1]
-  },
-]
-
-/*const MY_COLS: string[] = [
-  "Malfunction",
-  "Troleibus",
-  "Tramvai",
-  "Electrobus",
-  "Avtobus"
-]*/
-const MY_COLS: string[] = [
-  "Несправність",
-  "Тролейбус T3L",
-  "Трамвай",
-  "Електробус",
-  "Автобус"
-]
 
 @Component({
   selector: 'malfunction-report',
@@ -67,7 +28,6 @@ export class MalfunctionReportComponent implements OnInit {
       private statisticsService: StatisticsService,
       private vehicleTypeService: VehicleTypeService
   ) {
-    
   }
 
   ngOnInit() {
@@ -78,10 +38,9 @@ export class MalfunctionReportComponent implements OnInit {
         this.displayedColumns.push(vType.name);
       });
     });
-    //this.displayedColumns = MY_COLS;
+
     this.statisticsService.GetAllMalfunctionsStatistics(this.malfunctionSubgroupName).subscribe(data => {
       let rows = [];
-
       data.forEach(row => {
         rows.push(CreateMatTableRowFromStatistics(row, this.displayedColumns));
       });
@@ -90,16 +49,6 @@ export class MalfunctionReportComponent implements OnInit {
       this.dataSource.paginator = this.paginator;
       this.dataSource.sort = this.sort;
     });
-    
-    /*let rows = [];
-
-    MY_ROWS.forEach(row => {
-      rows.push(CreateMatTableRowFromStatistics(row, this.displayedColumns));
-    });
-
-    this.dataSource = new MatTableDataSource(rows);
-    this.dataSource.paginator = this.paginator;
-    this.dataSource.sort = this.sort;*/
   }
 
   applyFilter(filterValue: string) {
